Update recommendations and relist vaft
"Video Ad-Block, for Twitch (fork)" has been relisted in the recommendations as it does still work in terms of hiding the ads.

"vaft" has been improved and relisted. This solution is now a hybrid of embed / proxy / the "commercial break in progress" screen. If one method fails it moves to the next.